Переработали сортировку предложений
