+++ src/components/main/main.jsx
@@ -2,20 +2,11 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import {connect} from 'react-redux';
 
-import withSortingType from '../../hocs/with-sorting-type';
-import withOffersList from '../../hocs/with-offers-list';
-import SortingType from '../sorting-type/sorting-type';
 import CitiesList from '../cities-list/cities-list';
-import OffersList from '../offers-list/offers-list';
+import OffersContainer from '../offers-container/offers-container';
 import UserInfo from '../user-info/user-info';
-import CityMap from '../city-map/city-map';
 
-import offerPropTypes from '../../mocks/offer-prop-types';
-
-const ExtendedSortingType = withSortingType(SortingType);
-const ExtendedOffersList = withOffersList(OffersList);
-
-const Main = ({cities, city, cityOffers}) => {
+const Main = ({cities, city}) => {
   return (
     <div className="page page--gray page--main">
       <header className="header">
@@ -39,7 +30,7 @@ const Main = ({cities, city, cityOffers}) => {
       <main className="page__main page__main--index">
         <h1 className="visually-hidden">Cities</h1>
         <CitiesList cities={cities} currentCity={city} />
-        <ExtendedOffersList />
+        <OffersContainer city={city} />
       </main>
     </div>
   );
@@ -48,10 +39,9 @@ const Main = ({cities, city, cityOffers}) => {
 Main.propTypes = {
   city: PropTypes.string.isRequired,
   cities: PropTypes.arrayOf(PropTypes.string).isRequired,
-  cityOffers: PropTypes.arrayOf(offerPropTypes.isRequired).isRequired,
 };
 
-const mapStateToProps = (state) => ({city: state.city.name, cityOffers: state.cityOffers});
+const mapStateToProps = (state) => ({city: state.city.name});
 
 export {Main};
 export default connect(mapStateToProps)(Main);
